Web Development/am-gaap-event-journal-netsuite-tie-app/backend/requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
attrs
blinker
Brotli
cachetools
certifi
cffi
charset-normalizer
click
colorama
cryptography
Deprecated
et_xmlfile
Flask==3.0.3
Flask-Cors
Flask-JWT-Extended
Flask-Mail
flatbuffers==23.5.26
frozenlist
gast==0.5.4
google-api-core
google-api-python-client==2.134.0
google-auth==2.25.1
google-auth-httplib2
google-auth-oauthlib==1.1.0
google-cloud==0.34.0
google-cloud-bigquery==3.29.0
google-cloud-bigquery-datatransfer==3.15.3
google-cloud-core
google-cloud-pubsub==2.27.1
google-cloud-secret-manager==2.22.1
google-cloud-storage==2.14.0
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
grpc-google-iam-v1==0.13.1
grpcio
grpcio-status==1.63.0rc1
gunicorn==23.0.0
h2
hpack
httplib2
hyperframe
idna
importlib_metadata
importlib_resources
itsdangerous
Jinja2
jsonpointer==2.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
libcst
Markdown==3.5.1
MarkupSafe
multidict
mypy_extensions
numpy==2.2.3
oauthlib
openpyxl
opentelemetry-api
opentelemetry-sdk
opentelemetry-semantic-conventions
packaging
pandas==2.2.3
pkgutil_resolve_name
ply
propcache
proto-plus
protobuf==4.25.6
pyasn1
pyasn1_modules
pycparser
PyJWT
pyOpenSSL
pyparsing
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip
PySocks
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
pytz==2025.1
pyu2f
PyYAML
referencing
regex==2024.11.6
requests
requests-oauthlib
rpds-py
rsa
sip
six
toml
tomli
typing_extensions
typing_inspect
tzdata==2025.1
uritemplate
urllib3==1.26.12
Werkzeug==3.0.3
win_inet_pton
wrapt
XlsxWriter
xlwings==0.33.9
yarl
zipp
zstandard==0.23.0
